renamed the new project manager and feature builder widget files
