+++ Cognitive3DTest/Plugins/Cognitive3D/Source/Cognitive3DEditor/Private/Cognitive3DEditorModule.h
@@ -14,8 +14,8 @@
 #include "Editor/EditorStyle/Public/EditorStyleSet.h"
 #include "SceneSetupWidget.h"
 #include "ProjectSetupWidget.h"
-#include "SProjectManagerWidget.h"
-#include "SFeatureBuilderWidget.h"
+#include "ProjectManagerWidget.h"
+#include "FeatureBuilderWidget.h"
 #include "DynamicObjectManagerWidget.h"
 #include "WorkspaceMenuStructure.h"
 #include "CognitiveSettingsCustomization.h"
