+++ Cognitive3DTest/Plugins/Cognitive3D/Source/Cognitive3DEditor/Private/FeatureBuilderWidget.cpp
@@ -1,7 +1,7 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 
-#include "SFeatureBuilderWidget.h"
+#include "FeatureBuilderWidget.h"
 #include "SlateOptMacros.h"
 #include "Widgets/Layout/SScrollBox.h"
 #include "Widgets/Images/SImage.h"
